remove intermediate 'modules' subpackage
