+++ test/tests.conf
@@ -19,15 +19,15 @@ object: AS37271
 name: test
 
 [query-classes]
-native: rptk.modules.query.native.NativeQuery
-bgpq3: rptk.modules.query.bgpq3.Bgpq3Query
+native: rptk.query.native.NativeQuery
+bgpq3: rptk.query.bgpq3.Bgpq3Query
 
 [format-classes]
-test: rptk.modules.format.test.TestFormat
-json: rptk.modules.format.jsonf.JsonFormat
-yaml: rptk.modules.format.yamlf.YamlFormat
-plain: rptk.modules.format.plain.PlainFormat
-ios: rptk.modules.format.ios.IosFormat
-ios_null: rptk.modules.format.ios_null.IosNullFormat
-junos: rptk.modules.format.junos.JunosFormat
-bird: rptk.modules.format.bird.BirdFormat
+test: rptk.format.test.TestFormat
+json: rptk.format.jsonf.JsonFormat
+yaml: rptk.format.yamlf.YamlFormat
+plain: rptk.format.plain.PlainFormat
+ios: rptk.format.ios.IosFormat
+ios_null: rptk.format.ios_null.IosNullFormat
+junos: rptk.format.junos.JunosFormat
+bird: rptk.format.bird.BirdFormat
